+++ lib/td_dd/loader/metadata.ex
@@ -6,12 +6,40 @@ defmodule TdDd.Loader.Metadata do
 
   alias Ecto.Multi
   alias TdDd.DataStructures
+  alias TdDd.DataStructures.DataStructure
   alias TdDd.DataStructures.StructureMetadata
   alias TdDd.Repo
+  alias TdDd.Systems.System
 
   require Logger
 
+  @type metadata_record :: %{external_id: binary(), mutable_metadata: map()}
+
   @chunk_size 1_000
+  @unnest_external_ids "select external_id from unnest(?::text[]) t (external_id)"
+  @unnest_records "select * from unnest(?::jsonb[]) t (data)"
+  @expand_metadata "select data->>'external_id' as external_id, data->'mutable_metadata' as fields from data"
+
+  @doc """
+  Identifies external_ids which do not exist in the specified system
+  """
+  @spec missing_external_ids(atom, map, [metadata_record], System.t()) ::
+          {:error, [integer()]} | {:ok, []}
+  def missing_external_ids(_repo, _changes, records, %{id: system_id} = _system) do
+    case missing_external_ids(records, system_id) do
+      [] -> {:ok, []}
+      [_ | _] = ids -> {:error, ids}
+    end
+  end
+
+  @spec missing_external_ids([metadata_record], integer()) :: [integer()]
+  def missing_external_ids(records, system_id) do
+    records
+    |> Enum.map(& &1.external_id)
+    |> Enum.uniq()
+    |> Enum.chunk_every(@chunk_size)
+    |> Enum.flat_map(&select_missing_external_ids(system_id, &1))
+  end
 
   @doc """
   Logically delete metadata versions whose external_id is absent from the
@@ -35,6 +63,32 @@ defmodule TdDd.Loader.Metadata do
     |> do_update(ts)
   end
 
+  @spec merge_metadata(atom, map, [map], DateTime.t()) :: {:error, any} | {:ok, [integer()]}
+  def merge_metadata(_repo, %{} = _changes, records, ts), do: merge_metadata(records, ts)
+
+  @spec merge_metadata([map], DateTime.t()) :: {:error, any} | {:ok, [integer()]}
+  def merge_metadata([], _), do: {:ok, []}
+
+  def merge_metadata([_ | _] = records, ts) do
+    records
+    |> Enum.map(&cast_merge/1)
+    |> Enum.flat_map(fn
+      {:ok, res} -> [res]
+      {:error, _} -> []
+    end)
+    |> merge_existing_fields()
+    |> do_update(ts)
+  end
+
+  def cast_merge(%{} = params) do
+    import Ecto.Changeset
+
+    {%{}, %{external_id: :string, mutable_metadata: :map}}
+    |> cast(params, [:external_id, :mutable_metadata], empty_values: ["", %{}])
+    |> validate_required([:external_id, :mutable_metadata])
+    |> apply_action(:update)
+  end
+
   defp mutable_metadata_entry(%{external_id: external_id, mutable_metadata: %{} = mm})
        when mm != %{} do
     {external_id, mm}
@@ -44,6 +98,8 @@ defmodule TdDd.Loader.Metadata do
     {external_id, nil}
   end
 
+  defp do_update(%{} = entries, _ts) when entries == %{}, do: {:ok, []}
+
   defp do_update(%{} = entries_by_external_id, ts) do
     entries_by_external_id
     |> Enum.chunk_every(@chunk_size)
@@ -51,6 +107,31 @@ defmodule TdDd.Loader.Metadata do
     |> Enum.reduce(Multi.new(), &bulk_operations(&1, &2, ts))
     |> Repo.transaction()
     |> transform_result()
+  end
+
+  def merge_existing_fields(entries) do
+    entries
+    |> Enum.chunk_every(@chunk_size)
+    |> Enum.flat_map(&do_merge_existing_fields/1)
+    |> Map.new()
+  end
+
+  def do_merge_existing_fields(entries) do
+    "fields_to_merge"
+    |> with_cte("data", as: fragment(@unnest_records, ^entries))
+    |> with_cte("fields_to_merge", as: fragment(@expand_metadata))
+    |> join(:left, [rec], ds in DataStructure, on: ds.external_id == rec.external_id)
+    |> join(:left, [_rec, ds], sm in assoc(ds, :metadata_versions))
+    |> where([_rec, _ds, sm], is_nil(sm.deleted_at))
+    |> select([rec, _, sm], %{
+      external_id: rec.external_id,
+      existing_metadata: sm.fields,
+      merged_metadata: fragment("coalesce(?, '{}'::jsonb) || ?", sm.fields, rec.fields)
+    })
+    |> subquery()
+    |> where([e], is_nil(e.existing_metadata) or e.merged_metadata != e.existing_metadata)
+    |> select([e], {e.external_id, e.merged_metadata})
+    |> Repo.all()
   end
 
   defp transform_result({:error, failed_operation, _failed_value, _changes_so_far}) do
@@ -68,11 +149,11 @@ defmodule TdDd.Loader.Metadata do
           Enum.map(metadata_version, & &1.data_structure_id)
       end)
 
-    {:ok, structure_ids}
+    {:ok, Enum.uniq(structure_ids)}
   end
 
   defp bulk_operations({chunk, chunk_id}, multi, ts) do
-    structures_by_external_id =
+    metadata_by_external_id =
       chunk
       |> Enum.map(&elem(&1, 0))
       |> DataStructures.get_latest_metadata_by_external_ids()
@@ -80,8 +161,9 @@ defmodule TdDd.Loader.Metadata do
 
     chunk
     |> Enum.flat_map(fn {external_id, entry} ->
-      current_structure = Map.get(structures_by_external_id, external_id)
-      operations(current_structure, entry, ts)
+      metadata_by_external_id
+      |> Map.get(external_id)
+      |> operations(entry, ts)
     end)
     |> Enum.group_by(& &1.operation, &Map.delete(&1, :operation))
     |> Enum.reduce(multi, &reduce_multi(&1, &2, chunk_id, ts))
@@ -164,5 +246,17 @@ defmodule TdDd.Loader.Metadata do
     |> where([sm], is_nil(sm.deleted_at))
     |> where([sm], sm.data_structure_id in ^ids)
     |> Repo.update_all(set: [deleted_at: ts])
+  end
+
+  defp select_missing_external_ids(system_id, external_ids) do
+    DataStructure
+    |> where(system_id: ^system_id)
+    |> select([ds], [:id, :external_id])
+    |> subquery()
+    |> with_cte("external_ids", as: fragment(@unnest_external_ids, ^external_ids))
+    |> join(:right, [ds], id in "external_ids", on: ds.external_id == id.external_id)
+    |> where([ds, _], is_nil(ds.id))
+    |> select([_, i], i.external_id)
+    |> Repo.all()
   end
 end